+++ Source/gamemenu.h
@@ -10,6 +10,7 @@ namespace devilution {
 void gamemenu_on();
 void gamemenu_off();
 void gamemenu_handle_previous();
+void gamemenu_exit_game(bool bActivate);
 void gamemenu_quit_game(bool bActivate);
 void gamemenu_load_game(bool bActivate);
 void gamemenu_save_game(bool bActivate);
